+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Add new icons from patch 9.0.2
 
+### Fixed
+
+- Fix an error when using `-` (dash) as a keybind [#30]
+
 ## [0.11.1] - 2020-11-18
 
 ### Added
@@ -403,6 +407,7 @@
 [0.4.0]: https://github.com/Snakybo/Clicked/releases/tag/0.4.0
 [0.3.0]: https://github.com/Snakybo/Clicked/releases/tag/0.3.0
 
+[#30]: https://github.com/Snakybo/Clicked/issues/30
 [#29]: https://github.com/Snakybo/Clicked/issues/29
 [#27]: https://github.com/Snakybo/Clicked/issues/27
 [#26]: https://github.com/Snakybo/Clicked/issues/26